cell_tracker/requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
torch>=1.9.0
cellSAM @ git+https://github.com/vanvalenlab/cellSAM.git
scikit-image>=0.19.0
PyYAML>=5.4.0
tqdm>=4.65.0 